backend/requirements.txt
asyncio
websockets
mediapipe
opencv-python
pillow
numpy